// src/api/schema/category.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI spec version: 0.0.0
 */
import type { CategoryType } from './categoryType';

export interface Category {
  readonly id: number;
  /** @maxLength 100 */
  name: string;
  readonly types: readonly CategoryType[];
  readonly created_at: string;
  readonly updated_at: string;
}